Check all updates

+++ tgbot4j-core/src/main/java/ru/loolzaaa/tgbot4j/core/api/types/DirectMessagesTopic.java
@@ -14,10 +14,14 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class DirectMessagesTopic {
     /**
-     * Unique identifier of the topic
+     * Unique identifier of the topic. This number may have more
+     * than 32 significant bits and some programming languages
+     * may have difficulty/silent defects in interpreting it.
+     * But it has at most 52 significant bits, so a 64-bit integer
+     * or double-precision float type are safe for storing this identifier.
      */
     @JsonProperty("topic_id")
-    private Integer topicId;
+    private Long topicId;
 
     /**
      * Optional. Information about the user that created the topic.
